refactor EmoteCardModels
