+++ mod/twitch/src/main/java/tv/twitch/android/api/parsers/EmoteCardModelParser.java
@@ -1,8 +1,5 @@
 package tv.twitch.android.api.parsers;
 
-import kotlin.jvm.internal.BTTVDefaultConstructorMarker;
-import tv.twitch.android.provider.chat.model.EmoteCardModel;
-
 public class EmoteCardModelParser {
 
 }
