+++ mod/twitch/src/main/java/tv/twitch/android/shared/chat/emotecard/EmoteCardUiModel.java
@@ -1,6 +1,6 @@
 package tv.twitch.android.shared.chat.emotecard;
 
-import tv.twitch.android.provider.chat.model.EmoteCardModel;
+import tv.twitch.android.models.emotes.EmoteCardModel;
 
 public abstract class EmoteCardUiModel {
     public abstract String getEmoteUrl();
